Add files via upload
CCSE Clubs is a website for students of the College of Computer Science and Engineering (CCSE) at the University of Jeddah. It aims, to create a platform that connects all student clubs in the CCSE, which contributes to improving the quality of university activities for college students and providing them with better opportunities to invest in these creative energies.